master polish

+++ client/src/main/java/complatablefuture/client/app/ChangeController.java
@@ -8,6 +8,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @RestController
@@ -18,7 +19,7 @@ class ChangeController {
     private final ChangeService changeService;
 
     @GetMapping("/{amount}/{currencies}")
-    List<Change> changeCurrency(@PathVariable Integer amount ,
+    List<Change> changeCurrency(@PathVariable BigDecimal amount ,
                                 @PathVariable List<String> currencies) {
         return changeService.change(amount,currencies);
     }
