+++ client/src/main/java/complatablefuture/client/domain/model/Rate.java
@@ -3,9 +3,11 @@ package complatablefuture.client.domain.model;
 import lombok.Builder;
 import lombok.Value;
 
+import java.math.BigDecimal;
+
 @Value
 @Builder
 public class Rate {
     String code;
-    Double rate;
+    BigDecimal rate;
 }
